Widen branch_c64_t to long double, fixing overflow at the delegate boundary
Found via a mqt-fixed cross-check between dlx4sop-branch and
dlx4sop-treewidth: sop-solve --backend branch printed literal
amplitude_re/im: inf/-nan for pricingcall_indep_qiskit_17 (norm_h=2192),
while --backend treewidth on the identical qsop file produced a correct,
finite (if huge, ~1e327) value.

Root cause: branch_c64_t, the type branch.c's single-fourier machinery
uses to combine amplitudes, stored re/im as plain double. Components
delegated to treewidth/rankwidth come back as qsop_amplitude_t (long
double, matching their own raw pre-normalization magnitudes of roughly
2**(norm_h/2) -- routinely past a double's ~1.8e308 range once norm_h
exceeds ~2048). branch_try_single_mode_delegate narrowed the delegated
result into a branch_c64_t immediately on receipt, silently overflowing
to inf before any further combination happened.

Widened branch_c64_t's fields to long double (private to this file, no
header/ABI exposure) and updated its handful of helpers accordingly; the
one delegate-boundary narrowing cast is now a plain (lossless) copy. Left
the phase cache's own storage as double -- roots of unity are always
magnitude 1, so narrowing there is precision-only, never an overflow
risk -- and switched c64_accum_error's epsilon from DBL_EPSILON to
LDBL_EPSILON to match the now-actual arithmetic precision, mirroring
treewidth.c/rankwidth.c's existing convention.

Verified: branch now matches treewidth to full long-double precision on
the previously-inf/nan case; full test suite (53/53, including the
differential cross-backend check) still green.

+++ src/solve/branch.c
@@ -2646,17 +2646,24 @@ static void branch_root_of_unity(uint64_t r, uint32_t target_mode, uint64_t k, l
   *im = sinl(angle);
 }
 
+/* re/im are long double, not double: components delegated to treewidth/rankwidth
+ * come back as qsop_amplitude_t (long double), with raw (pre-normalization)
+ * magnitudes on the order of 2**(norm_h/2) -- routinely far beyond a double's
+ * ~1.8e308 range on deep circuits (norm_h > ~2048 is enough). Narrowing here
+ * would silently overflow to inf/nan before any combination even happens; see
+ * branch_try_single_mode_delegate below, the one place a delegated amplitude
+ * enters this type. */
 typedef struct branch_c64 {
-  double re;
-  double im;
+  long double re;
+  long double im;
 } branch_c64_t;
 
 static inline branch_c64_t c64_zero(void) {
-  return (branch_c64_t){0.0, 0.0};
+  return (branch_c64_t){0.0L, 0.0L};
 }
 
 static inline branch_c64_t c64_one(void) {
-  return (branch_c64_t){1.0, 0.0};
+  return (branch_c64_t){1.0L, 0.0L};
 }
 
 static inline branch_c64_t c64_add(branch_c64_t a, branch_c64_t b) {
@@ -2669,7 +2676,7 @@ static inline branch_c64_t c64_mul(branch_c64_t a, branch_c64_t b) {
 
 static inline void c64_accum_error(uint64_t ops, long double *err) {
   if (err != NULL) {
-    *err += (long double)ops * 8.0L * DBL_EPSILON;
+    *err += (long double)ops * 8.0L * LDBL_EPSILON;
   }
 }
 
@@ -2855,8 +2862,11 @@ static branch_c64_t branch_phase_lookup(branch_phase_cache_t *cache, uint64_t re
   }
   cache->used[idx] = 1U;
   cache->keys[idx] = residue;
-  cache->re[idx] = value.re;
-  cache->im[idx] = value.im;
+  /* The cache itself stores double-precision phases (magnitude always 1, so this
+   * narrowing never risks overflow) -- only the amplitude-carrying branch_c64_t
+   * values need long double range. */
+  cache->re[idx] = (double)value.re;
+  cache->im[idx] = (double)value.im;
   cache->len++;
   return value;
 }
@@ -3451,7 +3461,7 @@ static bool branch_try_single_mode_delegate(qsop_residual_t *residual,
     return ok;
   }
 
-  branch_c64_t amp = {(double)delegated_amp.re, (double)delegated_amp.im};
+  branch_c64_t amp = {delegated_amp.re, delegated_amp.im};
   const branch_c64_t constant =
       branch_phase_lookup(&state->phase_cache, qsop_residual_constant(residual));
   *out = c64_mul(constant, amp);
